refactor(fsi): Move hiddenfile and linkfile into fsi subpackages
Based on https://github.com/qri-io/qri/commit/4cb69f0f4ff8ad466a3b65d301761e507b46ee23. After discussing, we decided to put these packages under fsi/ instead of merging them into that directory.
